% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psvm4r.R
\docType{methods}
\name{initialize,LinearKernel-method}
\alias{initialize,LinearKernel-method}
\title{Create a LinearKernel. No parameters are necessary.}
\usage{
\S4method{initialize}{LinearKernel}(.Object, ...)
}
\description{
Create a LinearKernel. No parameters are necessary.
}
