% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psvm4r.R
\docType{methods}
\name{initialize,OneAgainstAllCSVC-method}
\alias{initialize,OneAgainstAllCSVC-method}
\title{Create a One-against-all C-SVC with the given parameters}
\usage{
\S4method{initialize}{OneAgainstAllCSVC}(.Object, num_classes,
  num_dimensions, kernel)
}
\arguments{
\item{num_classes}{number of distinct classes in the label vector of the problem}

\item{num_dimensions}{number of dimensions in the attributes table of the problem}

\item{kernel}{kernel object to use}
}
\description{
Create a One-against-all C-SVC with the given parameters
}
