% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psvm4r.R
\docType{methods}
\name{initialize,PolynomialKernel-method}
\alias{initialize,PolynomialKernel-method}
\title{Create a PolynomialKernel with the given parameters.}
\usage{
\S4method{initialize}{PolynomialKernel}(.Object, ...)
}
\arguments{
\item{degree}{degree of the exponent}

\item{constant}{additive constant}
}
\description{
Create a PolynomialKernel with the given parameters.
}
