% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psvm4r.R
\docType{methods}
\name{initialize,RbfKernel-method}
\alias{initialize,RbfKernel-method}
\title{Create a RbfKernel with the given parameter.}
\usage{
\S4method{initialize}{RbfKernel}(.Object, ...)
}
\arguments{
\item{gamma}{Parameter to control the radius. Note that gamma=1/(2*sigma^2) so a higher gamma yields a smaller radius}
}
\description{
Create a RbfKernel with the given parameter.
}
